recipe modal restructure
