+++ client/src/App.jsx
@@ -4,12 +4,11 @@ import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import Recipes from "./pages/Recipes.jsx"
 import Calendar from "./pages/Calendar.jsx"
 import CalendarMonth from "./pages/CalendarMonth.jsx"
-import RecipeModal from './components/RecipeModal.jsx';
+import RecipeModal from './components/RecipeModal/RecipeModal.jsx';
 import Home from "./pages/Home.jsx"
 import Navbar from './components/Navbar.jsx';
 import CreateUser from './pages/CreateUser.jsx';
 import Profile from './components/Profile/Profile.jsx';
-import TEST from './components/TEST.jsx';
 import { Link } from 'react-router-dom';
 import { RecipeContext } from "./ContextProvider"
 import EditProfile from './pages/EditProfile.jsx';
@@ -146,7 +145,6 @@ function App() {
           <Route path='/recipes' element={<Recipes setSelectedRecipe={setSelectedRecipe} setIsRecipeModal={setIsRecipeModal} setIsRecipeModalAdd={setIsRecipeModalAdd} />} />
           <Route path='/calendar' element={<Calendar isRecipeModal={isRecipeModal} setRecipes={setRecipes} setSelectedRecipe={setSelectedRecipe} setIsRecipeModal={setIsRecipeModal} setIsRecipeModalAdd={setIsRecipeModalAdd} calendar={calendar} setCalendar={setCalendar}/>} />
           <Route path='/calendar-month' element={<CalendarMonth isRecipeModal={isRecipeModal} setRecipes={setRecipes} setSelectedRecipe={setSelectedRecipe} setIsRecipeModal={setIsRecipeModal} setIsRecipeModalAdd={setIsRecipeModalAdd} calendar={calendar} setCalendar={setCalendar} />} />
-          <Route path='/test' element={<TEST />} />
         </Routes>
         {isRecipeModal && <RecipeModal isRecipeModal={isRecipeModal} isRecipeModalAdd={isRecipeModalAdd} setIsRecipeModal={setIsRecipeModal} selectedRecipe={selectedRecipe} calendar={calendar} setCalendar={setCalendar}/>}
         <footer className='footer'>
