+++ client/src/components/RecipeModal/RecipeModal.jsx
@@ -1,8 +1,12 @@
 import React, { useState, useEffect, useContext } from 'react';
-import '../css/recipemodal.css';
-import Draggable from "react-draggable";
-import CalendarModal from './CalendarModal/CalendarModal';
-import { RecipeContext } from '../ContextProvider';
+import '../../css/recipemodal.css';
+import CalendarModal from '../CalendarModal/CalendarModal';
+import { RecipeContext } from '../../ContextProvider';
+import Header from './Header';
+import Description from './Description';
+import Details from './Details';
+import CalendarDetails from './CalendarDetails';
+import Footer from './Footer';
 
 const RecipeModal = ({ isRecipeModalAdd, setIsRecipeModal, selectedRecipe, calendar, setCalendar }) => {
   const [date, setDate] = useState(new Date().toISOString().slice(0, 10));
@@ -16,44 +20,13 @@ const RecipeModal = ({ isRecipeModalAdd, setIsRecipeModal, selectedRecipe, calen
   const [adjustedVitaminA, setAdjustedVitaminA] = useState(0)
   const [adjustedVitaminC, setAdjustedVitaminC] = useState(0)
   const [adjustedIngredients, setAdjustedIngredients] = useState([]);
-
-  const { user, setUser } = useContext(RecipeContext)
-
-  //calendarmodal
   const [isCalendarModal, setIsCalendarModal] = useState(false)
 
-  const handleCalendarMouseOut = () => {
-    setIsCalendarModal(false)
-  }
-
-  const handleCalendarMouseOver = () => {
-    setIsCalendarModal(true)
-  }
-  //calendarmodal
-
+  const { user, setUser } = useContext(RecipeContext)
 
   const handleCloseModal = () => {
     setIsRecipeModal(false);
   };
-
-  const handleCheckChange = (e) => {
-    const curMealType = e.target.name
-    let updatedMealTypes = [...mealTypes]
-    if (e.target.checked) {
-      updatedMealTypes.push(curMealType)
-    } else {
-      updatedMealTypes = updatedMealTypes.filter(m => m !== curMealType)
-    }
-
-    updatedMealTypes = updatedMealTypes.sort((a, b) => {
-      if (a === "breakfast") return -1
-      if (a === "lunch" && b === "dinner") return -1
-      if (a === "lunch" && b === "breakfast") return 1
-      if (a === "dinner") return 1
-    })
-
-    setMealTypes(updatedMealTypes)
-  }
 
   const fetchPostNewRecipe = async (recipe) => {
     const userId = user.id
@@ -84,7 +57,6 @@ const RecipeModal = ({ isRecipeModalAdd, setIsRecipeModal, selectedRecipe, calen
       handleCloseModal();
     }
   };
-
 
   const checkAvailability = (recipe) => {
     let counter = servings
@@ -150,7 +122,6 @@ const RecipeModal = ({ isRecipeModalAdd, setIsRecipeModal, selectedRecipe, calen
     return true
   }
 
-
   const updateCalendar = (recipe) => {
 
     let counter = servings
@@ -189,29 +160,6 @@ const RecipeModal = ({ isRecipeModalAdd, setIsRecipeModal, selectedRecipe, calen
     setCalendar(updatedCalendar)
   }
 
-
-  const handleAddRecipe = async () => {
-
-    //    if (!user) return
-    //  if (mealTypes.length === 0) return
-
-    let recipeToAdd = { ...selectedRecipe }
-    recipeToAdd.yield = servings
-    recipeToAdd["startDate"] = date
-    recipeToAdd["mealTypes"] = mealTypes
-    recipeToAdd["caloriesPerServing"] = Math.round(recipeToAdd.calories / servings);
-
-    if (checkAvailability(recipeToAdd)) {
-      const addedRecipeId = await fetchPostNewRecipe(recipeToAdd)
-      recipeToAdd["id"] = addedRecipeId
-      updateCalendar(recipeToAdd)
-    } else {
-      alert("One or more meals overlapping!")
-    }
-
-  }
-
-
   const handleEditRecipe = async () => {
     if (mealTypes.length === 0) return
 
@@ -228,6 +176,26 @@ const RecipeModal = ({ isRecipeModalAdd, setIsRecipeModal, selectedRecipe, calen
     }
 
   };
+
+  const handleAddRecipe = async () => {
+    if (!user) return
+    if (mealTypes.length === 0) return
+
+    let recipeToAdd = { ...selectedRecipe }
+    recipeToAdd.yield = servings
+    recipeToAdd["startDate"] = date
+    recipeToAdd["mealTypes"] = mealTypes
+    recipeToAdd["caloriesPerServing"] = Math.round(recipeToAdd.calories / servings);
+
+    if (checkAvailability(recipeToAdd)) {
+      const addedRecipeId = await fetchPostNewRecipe(recipeToAdd)
+      recipeToAdd["id"] = addedRecipeId
+      updateCalendar(recipeToAdd)
+    } else {
+      alert("One or more meals overlapping!")
+    }
+
+  }
 
   const fetchPatchEditedRecipe = async (recipe) => {
     try {
@@ -277,9 +245,6 @@ const RecipeModal = ({ isRecipeModalAdd, setIsRecipeModal, selectedRecipe, calen
     }
   };
 
-  const increaseServings = () => setServings(prev => prev + 1);
-  const decreaseServings = () => setServings(prev => prev > 1 ? prev - 1 : 1);
-
 
   const adjustDetails = () => {
     const baseServings = isRecipeModalAdd ? selectedRecipe.yield : selectedRecipe.servings; //Á
@@ -327,104 +292,33 @@ const RecipeModal = ({ isRecipeModalAdd, setIsRecipeModal, selectedRecipe, calen
     }
   }, [selectedRecipe, isRecipeModalAdd]);
 
-  //<Draggable cancel="input, .modal-close-button, label, .modal-footer, img, .modal-description, .details-container, .servings-container, .info-container" bounds={{ top: -200, left: -1300, right: 250, bottom: 100 }}>
-
   return (
     <div className="modal-overlay">
       <div className="modal-content">
-        <div className="modal-header">
-          <h2 className='modal-header-title'>{selectedRecipe.label}</h2>
-          <button onClick={handleCloseModal} className="modal-close-button">❌</button>
-        </div>
+        <Header handleCloseModal={handleCloseModal} selectedRecipe={selectedRecipe} />
         <div className="modal-body">
           <div className="modal-image">
             <img src={selectedRecipe.image} alt={selectedRecipe.label} onError={(e) => e.target.src = "/altfood.png"} />
           </div>
-          <div className="modal-description">
-            <div class='container-title'>
-              <strong>Descriptions</strong>
-            </div>
-            <div className='box-text-container'>
-              <div><strong>Source:</strong> {selectedRecipe.source}</div>
-              <div><strong>Cuisine:</strong> {selectedRecipe.cuisineType}</div>
-              <div><strong>Dish Type:</strong> {selectedRecipe.dishType}</div>
-              <div><strong>Vitamins:</strong> Vitamin A: {adjustedVitaminA.toFixed(2)}Î¼g, Vitamin C: {adjustedVitaminC.toFixed(2)}mg</div>
-              <div><strong>Diet Labels:</strong> {selectedRecipe.dietLabels.join(', ')}</div>
-              <div><strong>Health Labels:</strong> {selectedRecipe.healthLabels.join(', ')}</div>
-              <a className='full-recipe-button' href={selectedRecipe.url}>View Full Recipe</a>
-            </div>
-          </div>
-          <div className="details-container">
-            <div className='details'>
-              <div class='container-title'>
-                <strong>Meal Details</strong>
-              </div>
-              <div className='box-text-container'>
-                <h3> Calories: <span className='span'>{adjustedCalories} kcal ({Math.round(selectedRecipe.caloriesPerServing)} kcal / serving)</span></h3>
-                <h3>Preparation time: <span className='span'>{selectedRecipe.totalTime > 0 ? `${selectedRecipe.totalTime} mins` : 'No prep time'}</span></h3>
-                <div className='ingredients'>
-                  <h3><strong>Ingredients:</strong></h3>
-                  <ul>
-                    {adjustedIngredients.map((ingredient, index) => (
-                      <li key={index}>{ingredient}</li>
-                    ))}
-                  </ul>
-                </div>
-                <h3 className='nutrition'><strong>Nutrition:</strong> Fat: <span className='span'>{adjustedFat.toFixed(2)}g, Protein: {adjustedProtein.toFixed(2)}g, Carbs: {adjustedCarbs.toFixed(2)}g</span></h3>
-              </div>
-            </div>
-          </div>
-          <div className="info-container">
-            <div class='container-title'>
-              <strong>Calendar Details</strong>
-            </div>
-            <div className="servings-container">
-              <div className='info-label' >Servings:</div>
-              <div className="servings-control">
-                <button className="servings-button" onClick={decreaseServings}>-</button>
-                <input type="number" value={servings} readOnly />
-                <button className="servings-button" onClick={increaseServings}>+</button>
-              </div>
-            </div>
-            <div className="date-container">
-              <div className='info-label'>Starting Date:</div>
-              <div className='date-input'>
-                <input type="date" value={date} onChange={(e) => setDate(e.target.value)} />
-              </div>
-              <div className='calendar-modal-div' onMouseOver={handleCalendarMouseOver} onMouseOut={handleCalendarMouseOut}>
-                <img className='calendar-modal-icon' src='/calendar.png'></img>
-              </div>
-            </div>
-            <div className="meal-types-container">
-              <div className='info-label'>Meal Types:</div>
-              <div className="meal-type-control">
-                <div className='meal-type-tick'>
-                  <label htmlFor="breakfast">Breakfast</label>
-                  <input onChange={(e) => handleCheckChange(e)} type="checkbox" id="breakfast" name='breakfast' checked={mealTypes.includes('breakfast')} />
-                </div>
-                <div className='meal-type-tick'>
-                  <label htmlFor="lunch">Lunch</label>
-                  <input onChange={(e) => handleCheckChange(e)} type="checkbox" id="lunch" name='lunch' checked={mealTypes.includes('lunch')} />
-                </div>
-                <div className='meal-type-tick'>
-                  <label htmlFor="dinner">Dinner</label>
-                  <input onChange={(e) => handleCheckChange(e)} type="checkbox" id="dinner" name='dinner' checked={mealTypes.includes('dinner')} />
-                </div>
-              </div>
-            </div>
-          </div>
+          <Description selectedRecipe={selectedRecipe} adjustedVitaminA={adjustedVitaminA} adjustedVitaminC={adjustedVitaminC} />
+          <Details
+            adjustedCalories={adjustedCalories}
+            selectedRecipe={selectedRecipe}
+            adjustedIngredients={adjustedIngredients}
+            adjustedFat={adjustedFat}
+            adjustedProtein={adjustedProtein}
+            adjustedCarbs={adjustedCarbs} />
+          <CalendarDetails
+            servings={servings}
+            setServings={setServings}
+            date={date}
+            setDate={setDate}
+            setIsCalendarModal={setIsCalendarModal}
+            setMealTypes={setMealTypes}
+            mealTypes={mealTypes}
+          />
         </div>
-        <div className="modal-footer">
-          {isRecipeModalAdd ? (
-            <button className="modal-footer-Add" onClick={handleAddRecipe}>Add Recipe</button>)
-            : (
-              <div className='modal-footer-edit-delete'>
-                <button className="modal-footer-Edit" onClick={handleEditRecipe}>Edit Recipe</button>
-                <button className="modal-footer-Delete" onClick={fetchDeleteRecipe}>Delete Recipe</button>
-              </div>
-            )}
-        </div>
-
+        <Footer isRecipeModalAdd={isRecipeModalAdd} handleAddRecipe={handleAddRecipe} handleEditRecipe={handleEditRecipe} fetchDeleteRecipe={fetchDeleteRecipe} />
         {isCalendarModal &&
           <CalendarModal calendar={calendar} date={date} />
         }
